refactor(jazz-next): assign signature inference ownership
